Get rid of validators package Fixes #496
