+++ apps/web/src/types/permissions.ts
@@ -1,0 +1,7 @@
+export enum Permissions {
+  DeleteScenarios = "delete:scenarios",
+  AddScenarios = "add:scenarios",
+  ViewLab = "view:lab",
+  ViewAdmin = "view:admin",
+  EditScenarios = "edit:scenarios",
+}
